Make get_user happen only once during Init Chat. Move Logger to infra.
