+++ src/application/chat_service.hpp
@@ -7,7 +7,7 @@
 #include "mentions.hpp"
 #include "string.hpp"
 #include "enums.hpp"
-#include "logger.hpp"
+#include "infra/logger.hpp"
 #include <functional>
 #include <memory>
 #include <vector>
