+++ src/presentation/chat.cpp
@@ -57,7 +57,7 @@ Chat Chat::init(const Pubnub::String& publish_key, const Pubnub::String& subscri
     }
     catch (...)
     {
-        chat.create_user(user_id, ChatUserData());
+        chat.create_user_for_init_chat(user_id, ChatUserData());
     }
 
     if (config.store_user_activity_timestamps) {
@@ -250,6 +250,12 @@ void Pubnub::Chat::register_logger_callback(std::function<void(Pubnub::pn_log_le
 void Chat::store_user_activity_timestamp() const
 {
     this->user_service->store_user_activity_timestamp();
+}
+
+Pubnub::User Chat::create_user_for_init_chat(const Pubnub::String& user_id, const Pubnub::ChatUserData& user_data) const
+{
+    //During init we skip get_user during create_user, as we already checked that the user doesn't exist
+    return this->user_service->create_user(user_id, user_data, true);
 }
 
 #ifdef PN_CHAT_C_ABI
